+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- avifEncoder now has a speed setting
+- codec_aom only flushes encoder when necessary (avoids lost frame packets)
+
 ### Changed
 - Updated libaom to more recent SHA in aom.cmd
 - Tweaked AVIF_LOCAL_AOM settings to play nice with libaom's usage of CMake's option()
+- Minor cleanup
 
 ## [0.4.7] - 2019-11-11
 ### Changed